--- a/Common/Util/WorkerThread.cs
+++ b/Common/Util/WorkerThread.cs
@@ -68,7 +68,11 @@ namespace QuantConnect.Util
                     // pass, when the token gets cancelled
                 }
             })
-            { IsBackground = true, Name = "Isolator Thread" };
+            {
+                IsBackground = true,
+                Name = "Isolator Thread",
+                Priority = ThreadPriority.Highest
+            };
             _workerThread.Start();
         }