--- a/Engine/DataFeeds/WorkScheduling/WorkQueue.cs
+++ b/Engine/DataFeeds/WorkScheduling/WorkQueue.cs
@@ -15,6 +15,7 @@
 */
 
 using System.Collections.Generic;
+using System.Threading;
 
 namespace QuantConnect.Lean.Engine.DataFeeds.WorkScheduling
 {
@@ -23,11 +24,17 @@ namespace QuantConnect.Lean.Engine.DataFeeds.WorkScheduling
         private readonly List<WorkItem> _workQueue;
 
         /// <summary>
+        /// Event used to notify there is work ready to execute in this queue
+        /// </summary>
+        public AutoResetEvent WorkAvailableEvent { get; }
+
+        /// <summary>
         /// Creates a new instance
         /// </summary>
         public WorkQueue()
         {
             _workQueue = new List<WorkItem>();
+            WorkAvailableEvent = new AutoResetEvent(false);
         }
 
         /// <summary>
@@ -47,13 +54,22 @@ namespace QuantConnect.Lean.Engine.DataFeeds.WorkScheduling
         /// </summary>
         public void Sort()
         {
+            var notifyWork = false;
             lock (_workQueue)
             {
                 foreach (var item in _workQueue)
                 {
                     item.UpdateWeight();
+                    if (item.Weight < WeightedWorkScheduler.MaxWorkWeight)
+                    {
+                        notifyWork = true;
+                    }
                 }
                 _workQueue.Sort(WorkItem.Compare);
+            }
+            if (notifyWork)
+            {
+                WorkAvailableEvent.Set();
             }
         }
 
@@ -79,7 +95,7 @@ namespace QuantConnect.Lean.Engine.DataFeeds.WorkScheduling
             {
                 if (_workQueue.Count != 0)
                 {
-                    potentialWorkItem = _workQueue[0];
+                    potentialWorkItem = _workQueue[_workQueue.Count - 1];
 
                     // if the weight is at its maximum value return null
                     // this is useful to space out in time this work