… Engine/QuantConnect.Lean.Engine.csproj
… Tests/Engine/DataFeeds/SubscriptionUtilsTests.cs