<!--- Provide a general summary of your changes in the Title above -->

#### Description
<!--- Describe your changes in detail -->
- The new scheduler will create a dedicated thread pool
- Work will be prioritized based on their weight and new items will have
highest priority. The weight will be determined by the size of the
subscription enqueueable.
- The work queue will be sorted and the weights updated by a dedicated
thread
- There will be a maximum weight value that will 'disable' a work item
until their weight comes down
- The consumer will not know about workers or anything alike anymore.
- Applying a general max work queue size of 400 items, this will reduce
CPU and RAM usage. Can be set by config.

**Cloud performance tests**
MASTER
596.56 seconds at 2k data points per second. Processing total of 897,569 data points.
526.65 seconds at 2k data points per second. Processing total of 897,569 data points.
513.98 seconds at 2k data points per second. Processing total of 897,569 data points.
*Ram usage for the 3 backtests peaked at ~3GB*
PR
544.06 seconds at 2k data points per second. Processing total of 897,569 data points.
552.59 seconds at 2k data points per second. Processing total of 897,569 data points.
495.34 seconds at 2k data points per second. Processing total of 897,569 data points.
*Ram usage for the 3 backtests peaked at ~2.8GB*

<details><summary>Algorithm used for cloud performance tests</summary>
<p>

```
public class CoarseFineFundamentalRegressionAlgorithm : QCAlgorithm
{
    private const int NumberOfSymbolsCoarse = 500;
    private const int NumberOfSymbolsFine = 10;
    private readonly Dictionary<Symbol, Symbol> _symbols = new Dictionary<Symbol, Symbol>();

    public override void Initialize()
    {
        UniverseSettings.Resolution = Resolution.Daily;

        SetStartDate(2018, 1, 1);
        SetEndDate(2019, 1, 1);
        SetCash(1000000);

        AddUniverse(CoarseSelectionFunction, FineSelectionFunction);
        Schedule.On(DateRules.EveryDay(), TimeRules.Every(TimeSpan.FromDays(1)), PlotRam);
    }

    public void PlotRam()
    {
        Plot("RAM", "Managed", (double)GC.GetTotalMemory(false) / (1024 * 1024));
        Plot("RAM", "Total", (double)OS.ApplicationMemoryUsed);
    }

    public IEnumerable<Symbol> CoarseSelectionFunction(IEnumerable<CoarseFundamental> coarse)
    {
        var sortedByDollarVolume = coarse
            .Where(x => x.HasFundamentalData)
            .OrderByDescending(x => x.DollarVolume);

        var top5 = sortedByDollarVolume.Take(NumberOfSymbolsCoarse);

        return top5.Select(x => x.Symbol);
    }

    public IEnumerable<Symbol> FineSelectionFunction(IEnumerable<FineFundamental> fine)
    {
        var sortedByPeRatio = fine.OrderByDescending(x => x.ValuationRatios.PERatio);

        var topFine = sortedByPeRatio
            .Where(fundamental =>
                fundamental.AssetClassification.MorningstarIndustryGroupCode == MorningstarIndustryGroupCode.Biotechnology
                ||
                fundamental.AssetClassification.MorningstarIndustryGroupCode == MorningstarIndustryGroupCode.DrugManufacturers)
            .Take(NumberOfSymbolsFine)
            .Select(x => x.Symbol)
            .ToList();

        foreach (var symbol in topFine)
        {
            _symbols[symbol] = AddData<TiingoNews>(symbol).Symbol;
        }
        return topFine;
    }

    public override void OnSecuritiesChanged(SecurityChanges changes)
    {
        foreach (var changesRemovedSecurity in changes.RemovedSecurities)
        {
            if (_symbols.ContainsKey(changesRemovedSecurity.Symbol))
            {
                RemoveSecurity(_symbols[changesRemovedSecurity.Symbol]);
                _symbols.Remove(changesRemovedSecurity.Symbol);
            }
        }
    }
}
```

</p>
</details>

**Local windows performance tests were executed with:**
    - Max frequency limited. Boost frequency disabled. No thermal nor power limit throttling.
    - `Ticks` and `TradeBars` returning `SPY` symbol in their source

**Test 1:**
```
public override void Initialize()
{
    SetStartDate(2015, 11, 10);
    SetEndDate(2015, 11, 11);
    foreach (var symbol in Symbols.Equity.All.Take(50))
    {
        AddSecurity(SecurityType.Equity, symbol, Resolution.Tick);
    }
}
```
MASTER
43.46 seconds at 618k data points per second. Processing total of 26,841,371 data points.
45.00 seconds at 596k data points per second. Processing total of 26,841,371 data points.
43.83 seconds at 612k data points per second. Processing total of 26,841,371 data points.
Commit 1
42.81 seconds at 627k data points per second. Processing total of 26,841,371 data points.
43.13 seconds at 622k data points per second. Processing total of 26,841,371 data points.
42.48 seconds at 632k data points per second. Processing total of 26,841,371 data points.
Commit 2
42.39 seconds at 633k data points per second. Processing total of 26,841,371 data points.

**Test 2:**
```
public override void Initialize()
{
    SetStartDate(2015, 11, 10);
    SetEndDate(2015, 11, 11);
    foreach (var symbol in Symbols.Equity.All.Take(100))
    {
        AddSecurity(SecurityType.Equity, symbol, Resolution.Tick);
    }
}
```
MASTER 
93.60 seconds at 574k data points per second. Processing total of 53,682,738 data points.
92.75 seconds at 579k data points per second. Processing total of 53,682,738 data points.
93.49 seconds at 574k data points per second. Processing total of 53,682,738 data points.
Commit 1
86.74 seconds at 619k data points per second. Processing total of 53,682,738 data points.
88.27 seconds at 608k data points per second. Processing total of 53,682,738 data points.
89.48 seconds at 600k data points per second. Processing total of 53,682,738 data points.
Commit 2
88.70 seconds at 605k data points per second. Processing total of 53,682,738 data points.

**Test 3:**
```
public override void Initialize()
{
    SetStartDate(2015, 8, 10);
    SetEndDate(2015, 11, 11);
    foreach (var symbol in Symbols.Equity.All.Take(400))
    {
        AddSecurity(SecurityType.Equity, symbol, Resolution.Minute);
    }
}
```
MASTER
45.21 seconds at 197k data points per second. Processing total of 8,910,675 data points.
43.91 seconds at 203k data points per second. Processing total of 8,910,674 data points.
45.17 seconds at 197k data points per second. Processing total of 8,910,674 data points.
Commit 1
38.35 seconds at 232k data points per second. Processing total of 8,910,674 data points.
38.09 seconds at 234k data points per second. Processing total of 8,910,673 data points.
38.20 seconds at 233k data points per second. Processing total of 8,910,673 data points.
Commit 2
37.45 seconds at 238k data points per second. Processing total of 8,910,674 data points.

**Test 4:**
```
public override void Initialize()
{
    SetStartDate(2010, 1, 1);
    SetEndDate(2018, 1, 1);
    foreach (var symbol in Symbols.Equity.All.Take(400))
    {
        AddSecurity(SecurityType.Equity, symbol, Resolution.Daily);
    }
}
```
MASTER
7.43 seconds at 90k data points per second. Processing total of 669,511 data points.
7.22 seconds at 93k data points per second. Processing total of 669,511 data points.
7.23 seconds at 93k data points per second. Processing total of 669,479 data points.
Commit 1
6.28 seconds at 107k data points per second. Processing total of 669,543 data points.   
6.04 seconds at 111k data points per second. Processing total of 669,479 data points.
6.48 seconds at 103k data points per second. Processing total of 669,511 data points.
Commit 2
6.04 seconds at 111k data points per second. Processing total of 669,479 data points.

**Test 5:**
```
public override void Initialize()
{
    UniverseSettings.Resolution = Resolution.Daily;
    SetStartDate(2015, 1, 1);
    SetEndDate(2018, 1, 1);
    AddUniverse(CoarseSelectionFunction);
}
public static IEnumerable<Symbol> CoarseSelectionFunction(IEnumerable<CoarseFundamental> coarse)
{
    var sortedByDollarVolume = coarse.OrderByDescending(x => x.DollarVolume);
    var top = sortedByDollarVolume.Take(400);
    return top.Select(x => x.Symbol);
}
```
MASTER
46.02 seconds at 128k data points per second. Processing total of 5,894,398 data points.
45.07 seconds at 131k data points per second. Processing total of 5,894,390 data points.
45.06 seconds at 131k data points per second. Processing total of 5,894,374 data points.
Commit 1
31.99 seconds at 184k data points per second. Processing total of 5,894,374 data points.
31.51 seconds at 187k data points per second. Processing total of 5,894,395 data points.
31.39 seconds at 188k data points per second. Processing total of 5,894,410 data points.
Commit 2
31.15 seconds at 189k data points per second. Processing total of 5,894,384 data points.

**Test 6:**
```
public override void Initialize()
{
    UniverseSettings.Resolution = Resolution.Minute;
    SetStartDate(2017, 10, 1);
    SetEndDate(2018, 1, 1);
    AddUniverse(CoarseSelectionFunction);
}
public static IEnumerable<Symbol> CoarseSelectionFunction(IEnumerable<CoarseFundamental> coarse)
{
    var sortedByDollarVolume = coarse.OrderByDescending(x => x.DollarVolume);
    var top = sortedByDollarVolume.Take(400);
    return top.Select(x => x.Symbol);
}
```
MASTER
70.78 seconds at 145k data points per second. Processing total of 10,230,414 data points.
69.98 seconds at 146k data points per second. Processing total of 10,230,414 data points.
74.70 seconds at 137k data points per second. Processing total of 10,230,413 data points.
Commit 1
48.91 seconds at 209k data points per second. Processing total of 10,230,412 data points.
48.40 seconds at 211k data points per second. Processing total of 10,230,413 data points.
47.98 seconds at 213k data points per second. Processing total of 10,230,413 data points.
Commit 2
46.52 seconds at 220k data points per second. Processing total of 10,230,413 data points.
CLOUD MASTER - SetStartDate(2017, 9, 1);
94.91 seconds at 142k data points per second. Processing total of 13,475,195 data points.
92.07 seconds at 146k data points per second. Processing total of 13,475,195 data points.
CLOUD PR commit 2 - SetStartDate(2017, 9, 1);
58.69 seconds at 230k data points per second. Processing total of 13,475,195 data points.
53.69 seconds at 251k data points per second. Processing total of 13,475,195 data points.
**Test 7:**
```
public override void Initialize()
{
    SetStartDate(2017, 6, 1);
    SetEndDate(2018, 1, 1);
    foreach (var symbol in Symbols.Equity.All.Take(400))
    {
        AddData<TiingoNews>(AddSecurity(SecurityType.Equity, symbol, Resolution.Daily).Symbol);
    }
}
```
MASTER
63.61 seconds at 302k data points per second. Processing total of 19,195,228 data points.
64.15 seconds at 299k data points per second. Processing total of 19,195,225 data points.
65.52 seconds at 293k data points per second. Processing total of 19,195,223 data points.
Commit 1
58.06 seconds at 326k data points per second. Processing total of 18,907,289 data points.
58.07 seconds at 326k data points per second. Processing total of 18,907,289 data points.
58.84 seconds at 321k data points per second. Processing total of 18,907,289 data points.
Commit 2
57.11 seconds at 336k data points per second. Processing total of 19,195,223 data points.

**Test 8:**
```
HistoryRequestBenchmark
```
MASTER
22.08 seconds at 41k data points per second. Processing total of 907,031 data points.
22.08 seconds at 41k data points per second. Processing total of 907,031 data points.
Commit 1
26.09 seconds at 35k data points per second. Processing total of 907,031 data points.
25.69 seconds at 35k data points per second. Processing total of 907,031 data points.
Commit 2
21.68 seconds at 42k data points per second. Processing total of 907,031 data points.
21.69 seconds at 42k data points per second. Processing total of 907,031 data points.

#### Related Issue
<!--- This project only accepts pull requests related to open issues -->
<!--- If suggesting a new feature or change, please discuss it in an issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps to reproduce -->
<!--- Please link to the issue here: -->
Closes https://github.com/QuantConnect/Lean/issues/3976
#### Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
- Faster is better but efficient and faster is the sky
#### Requires Documentation Change
<!--- Please indicate if these changes will require updates to documentation, and if so, specify what changes are required -->
N/A
#### How Has This Been Tested?
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran to -->
<!--- see how your change affects other areas of the code, etc. -->
Unit and regression tests, performance tests.
#### Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] Refactor (non-breaking change which improves implementation)
- [x] Performance (non-breaking change which improves performance. Please add associated performance test and results)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)
- [ ] Non-functional change (xml comments/documentation/etc)

#### Checklist:
<!--- The following is a checklist of items that MUST be completed before a PR is accepted -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [x] My code follows the code style of this project.
- [x] I have read the **CONTRIBUTING** [document](https://github.com/QuantConnect/Lean/blob/master/CONTRIBUTING.md).
- [ ] I have added tests to cover my changes. <!--- If not applicable, please explain why -->
- [x] All new and existing tests passed.
- [x] My branch follows the naming convention `bug-<issue#>-<description>` or `feature-<issue#>-<description>`

<!--- Template inspired by https://www.talater.com/open-source-templates/#/page/99 -->