@jaredbroad 2020-01-10
Big one congrats @Martin-Molinero; leaving your mark at the core of LEAN =)